+++ sunpy/map/__init__.py
@@ -1,14 +1,10 @@
 """SunPy Maps"""
-__author__ = "Keith Hughitt"
-__email__ = "keith.hughitt@nasa.gov"
+from sunpy.map.mapbase import GenericMap  # isort:skip
 
-
-from sunpy.map.mapbase import GenericMap
-
-from . mapsequence import MapSequence
-from . compositemap import CompositeMap
-
-from sunpy.map.maputils import *
-from sunpy.map.maphelper import *
-from sunpy.map.map_factory import Map
 from sunpy.map import sources
+from sunpy.map.header_helper import *
+from sunpy.map.map_factory import Map
+from sunpy.map.maputils import *
+
+from .compositemap import CompositeMap
+from .mapsequence import MapSequence
