+++ sunpy/map/tests/test_header_helper.py
